Composition for Taskbar, Colors Cleaned, and Submit Form
Changed layout of components to be slightly more compositional.
Taskbar now exists on the app.js file, allowing for direct prop distribution instead of drilling.
Moved colors and themeing to a seperate file and folder, so that it doesn't clutter app.js more than it already is.
Users are now redirected to sign in page regardless of what page they go on while unauthorized.
Users are redirected to home page if they attempt to go to a page without any routing and are logged in.
Taskbar now calls useLocation to determine where bubble will go as well as whether it displays the back button.
Users can submit forms for login and sign up by pressing enter as form was encased in form component, go figure.