Description Editing and Styling Fixes
Edited ThingsIdChange
Added route to HabitController: {id}/description. Patch maps, so that you can update the description of the habit.
    Still has issues with errors regarding length, but I just forced the front end not to do that. There's definitely a better way.
    Added two queries to HabitRepository, updateDescription, and findHabitByUserAndId
Fixed minor styling problems with description text in popout on HabitBadgeProgress component.
Allowed front end to edit description by adding a text box that can be enabled by clicking the description. Not intuitive for mobile users, as they won't even get a pop up.
    Added tons of TODOs for this, requires a lot of work to get to a good spot.
    Had to pass user, habits, and setHabits to the component for this to work. In order to make this work more efficiently, should move to composing, and having global state for all relevant variables.
Changed background on the login pages to be background3. Looks nice.
Added a set of colors(fallbackText) to the colors.js palette. Allows you to set text based on dark or light mode that will stand out on the primary colors on papers and such.